Update fyfe to new template
A few issues/problems:

(1) The 3 long Bios push the abstract into the next page, which looks
bad.
Maybe we either need to (A) restrict the length of the bio to ~50 words,
or (B) push them to the end of the article??

(2) Did I do the multi Author symbol marking the way you want it?

(3) Citations do not appear correctly when I "build and view" the pdf
document from the .tex file.
I'm not sure if this is because I am doing something wrong, or if it is
because my binaries aren't updated to your latest ijsra version.

(4) Table 1 is automatically being inserted within  an "enumerate" list
(pg. 6), which looks bad... not sure how to change

(5) Table 2 has its own page, even though it takes up only 2/3 of a page
(p.7), which also looks bad

Let me know what else needs fixing here. Thanks for your patience!