Davis First Commit
This was just copy-pasting the draft into the template (the draft itself
contains my to-do list). I have a number of questions:
(1) So Gonzalo will put the references into the file (or the proper
format) for me to cite within the .tex file?
(2) The author includes ~6 footnotes... do we have a footnote
environment/command for teh template?
(3) The author provides no Biographical information. Should I ask
Gonzalo to have him send 2 sentances?
(4) There is a reference to a URL hyperlink. Is there a way to embed
this in LaTeX
(5) Mostof the author's citations include specific APA style page
numbers. Should I ignore these, or include them somehow?