#33 additional minor edits; see description
Notes for fixing
- quotation marks for small passages not proper (both right quote marks)
- possible to put original publication date and republished date? I
think Spivak 2010 is really Spivak 2010 [1983]? Author should confirm
whether or not this is the case.

- - - - -

Notes for author
- author needs to check over bib as it was noticed that some books were
not indicated that they were edited books
- author needs to check that Benson and Cook in-text citation as in the
proof is okay (3rd paragraph) [Word doc has this as the reference
"Ubelaker, cited in Benson and Cook 2001, 168"]
- source for Schliemann quote on page 3 of proof needed ('the face that
launched a thousand ships.')
- source for Weber quote on page 4 needed ('full understanding of the
facts')
- Is Spivak 2010 is really Spivak 2010 [1983]? I.e. was it republished
in a collection? Author should confirm whether or not this is the case.

- In references list but not used in-text:
=Emerson, T. et al. 2016. ‘Paradigms Lost: Reconfiguring Cahokia's Mound
72 Beaded Burial.’ American Antiquity 81(3), 405-25.
=Fowler, M. 1977. ‘Explorations into Cahokia Archaeology.’ Illinois
Archeological Survey 7, 18-26.
=Pauketat, T. and Emerson, T. 1997. Cahokia: Domination and Ideology in
the Mississippian World. Lincoln: University of Nebraska Press, 2, 101,
187, 237, 252-3, 255, 265.
=Thompson, A. Hedman, K. and Slater, P. 2015. ‘New Dental and Isotope
Evidence of Biological Distance and Place of Origin for Mass Burial
Groups at Cahokia’s Mound 72.’ American Journal of Physical Anthropology
158, 341-57.